Warning when use is incomplete
